Rename ner into tag
