+++ deeplima/apps/CMakeLists.txt
@@ -37,20 +37,20 @@ install(TARGETS deeplima-train-segm
 ############################################################
 # Train NER models
 
-add_executable(deeplima-train-ner
-  train-ner.cpp
+add_executable(deeplima-train-tag
+  train-tag.cpp
 )
 
-target_link_libraries(deeplima-train-ner
+target_link_libraries(deeplima-train-tag
   PRIVATE
-  train_ner
+  train_tag
   conllu
   Boost::program_options
   Boost::filesystem
   version
 )
 
-install(TARGETS deeplima-train-ner
+install(TARGETS deeplima-train-tag
   DESTINATION bin
   COMPONENT runtime
 )
