+++ deeplima/libs/convert/from_torch/CMakeLists.txt
@@ -33,7 +33,7 @@ add_library(convert_from_torch
 target_link_libraries(convert_from_torch
   PRIVATE
   birnn_classifier_for_segmentation
-  birnn_classifier_for_ner
+  birnn_classifier_for_tag
   model_lemmatization
   ${TORCH_LIBRARIES}
 )
@@ -48,7 +48,7 @@ if (NOT DEEPLIMA_STANDALONE)
   target_link_libraries(convert_from_torch-shared
     PRIVATE
     birnn_classifier_for_segmentation
-    birnn_classifier_for_ner
+    birnn_classifier_for_tag
     model_lemmatization
     ${TORCH_LIBRARIES}
   )
