+++ deeplima/libs/convert/from_torch/entity_tagging_model.cpp
@@ -18,7 +18,7 @@
 */
 
 #include "deeplima/eigen_wrp/tagging_eigen_inference_impl.h"
-#include "tasks/ner/model/birnn_classifier_for_ner.h"
+#include "tasks/tag/model/birnn_classifier_for_tag.h"
 
 #include "convert_from_torch.h"
